+++ src/main/kotlin/cn/luoym/bookreader/skylarkreader/action/NextPageAction.kt
@@ -1,10 +1,11 @@
 package cn.luoym.bookreader.skylarkreader.action
 
 import cn.luoym.bookreader.skylarkreader.extensions.Context
+import cn.luoym.bookreader.skylarkreader.utils.ReaderBundle
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 
-class NextPageAction: AnAction() {
+class NextPageAction: AnAction(ReaderBundle.message("skylark.reader.action.text.next.page")) {
 
     override fun actionPerformed(p0: AnActionEvent) {
         val context = Context.instance