+++ CHANGELOG.md
@@ -69,3 +69,8 @@
 
 - 新建项目
 - 通过[intellij-platform-plugin-template](https://github.com/JetBrains/intellij-platform-plugin-template)初始化项目
+- 实现了通过console面板阅读text电子书功能
+- 实现了epub电子书阅读功能
+- 实现了书架管理功能
+- 实现了通过status bar widget阅读text电子书功能
+- 添加了快捷键翻页、清除阅读面板内容